==> News Sources - Gathering, Cleaning, Combining/01 Data Gathering_Tribune.ipynb <==
import pandas as pd
import requests
from bs4 import BeautifulSoup
import time

# Scrape The Tribune India using BeautifulSoup

https://www.tribuneindia.com/archive.aspx

Extracting headline, date published, caption and url from a topic word search in India Tribune Archives from 2010 to 2019, for the five keywords/phrases below:


1) flood

2) flooded

3) flooding

4) monsoon

5) heavy rain

This was initally done in five seperate notebooks. 

# WARNING: this notebook will take a long time to run

The methodology behind this process was to first, extract as much surface level information as possible about the articles listed in the resultant archive search for each word. Then, using the urls found, the text from each article was extracted and stored in a seperate dataframe. Next the two data frames were merged and the superfluous columns were dropped. The resultant CSV had three columns, url, date and text. 

These csv's are combined in tribune_data2

## Flood

def web_df():
    
    #url is the first page that appears after clicking go when searching The Tribune India archives
    url = 'https://www.tribuneindia.com/archive_search.aspx?txt=Flood&catid=2&fd=01/01/2010%2012:00:00%20AM&td=11/01/2019%2012:00:00%20AM'
    res = requests.get(url)

    print(res.status_code)
    soup = BeautifulSoup(res.content, 'lxml')
    
    urls = [i('a', href=True)[0]['href'] for i in soup.find_all('div', {'class': 'listing'})]
    headlines = [i('a')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    dates = [i('span')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    captions = [i('p')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    
    d = {'headline': headlines, 
     'date': dates, 
     'caption': captions, 
     'url': urls}
    
    df_mother = pd.DataFrame(d)
    
    time.sleep(2)
    
    #change range according to how many pages available in The Tribune India archive for specified topic
    for i in range(2,16): 
        
        url1 = url + '&page=' + str(i)
        res = requests.get(url1)
        
        print(res.status_code)
        
        soup = BeautifulSoup(res.content, 'lxml')
        
        urls = [i('a', href=True)[0]['href'] for i in soup.find_all('div', {'class': 'listing'})]
        headlines = [i('a')[0].text for i in soup.find_all('div', {'class': 'listing'})]
        dates = [i('span')[0].text for i in soup.find_all('div', {'class': 'listing'})]
        captions = [i('p')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    
        d2 = {'headline': headlines, 
             'date': dates, 
             'caption': captions, 
             'url': urls}
    
        df = pd.DataFrame(d2)

        print(len(urls))
        print(len(dates))
        
        df_mother = pd.concat([df_mother, df], ignore_index=True)
    
        time.sleep(2)
        
#     return(col1, col2)
    return(df_mother)
        

url_df = web_df()

print(type(url_df))
print(url_df.shape)

url_df.to_csv('../data/trib_flood_url.csv', index=False)

df = pd.read_csv("../data/trib_flood_url.csv")

def content_scrape():
    
    content_dict = {}
    
    #range changes based on how many urls pulled
    for i in range(301):
        url_base = 'https://www.tribuneindia.com/'
        url = url_base + df['url'][i]

        res = requests.get(url)

        print(res.status_code)
        soup = BeautifulSoup(res.content, 'lxml')
        content = [i.text.replace('\n', ' ') for i in soup.find_all('span', {'class': 'storyText'})]
        
        content_dict.update({i: content})
        
        time.sleep(2)
    
    return content_dict
    

content_dict = content_scrape()

len(content_dict)

trib_flood_text = pd.DataFrame.from_dict(content_dict, orient='index')
trib_flood_text.rename(columns={0: 'text', inplace=True})
trib_flood_text.to_csv('../data/trib_flood_text.csv', index=False)

## Flooded

def web_df():

    url = 'https://www.tribuneindia.com/archive_search.aspx?txt=Flooded&catid=2&fd=01/01/2010%2012:00:00%20AM&td=11/01/2019%2012:00:00%20AM'
    res = requests.get(url)

    res.status_code
    soup = BeautifulSoup(res.content, 'lxml')
    
    urls = [i('a', href=True)[0]['href'] for i in soup.find_all('div', {'class': 'listing'})]
    headlines = [i('a')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    dates = [i('span')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    captions = [i('p')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    
    d = {'headline': headlines, 
     'date': dates, 
     'caption': captions, 
     'url': urls}
    
    df_mother = pd.DataFrame(d)
    
    time.sleep(2)
    
    #change range according to how many pages available in The Tribune India archive for specified topic
    for i in range(2,12): 
        
        url1 = url + '&page=' + str(i)
        res = requests.get(url1)
        
        print(res.status_code)
        
        soup = BeautifulSoup(res.content, 'lxml')
        
        urls = [i('a', href=True)[0]['href'] for i in soup.find_all('div', {'class': 'listing'})]
        headlines = [i('a')[0].text for i in soup.find_all('div', {'class': 'listing'})]
        dates = [i('span')[0].text for i in soup.find_all('div', {'class': 'listing'})]
        captions = [i('p')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    
        d2 = {'headline': headlines, 
             'date': dates, 
             'caption': captions, 
             'url': urls}
    
        df = pd.DataFrame(d2)

        print(len(urls))
        print(len(dates))
        
        df_mother = pd.concat([df_mother, df], ignore_index=True)
    
        time.sleep(2)

    return(df_mother)
        

url_df = web_df()

print(type(url_df))
print(url_df.shape)

url_df.to_csv('../data/trib_flooded_url.csv', index=False)

df = pd.read_csv("../data/trib_flooded_url.csv")

def content_scrape():
    
    content_dict = {}
    
    #range changes based on how many urls pulled
    for i in range(232):
        url_base = 'https://www.tribuneindia.com/'
        url = url_base + df['url'][i]

        res = requests.get(url)

        print(res.status_code)
        soup = BeautifulSoup(res.content, 'lxml')
        content = [i.text.replace('\n', ' ') for i in soup.find_all('span', {'class': 'storyText'})]
        
        content_dict.update({i: content})
        
        time.sleep(2)
    
    return content_dict
    

content_dict = content_scrape()

len(content_dict)

trib_flooded_text = pd.DataFrame.from_dict(content_dict, orient='index')
trib_flooded_text.rename(columns={0: 'text'}, inplace=True)
trib_flooded_text.to_csv('../data/trib_flooded_text.csv', index=False)

## Flooding

def web_df():

    url = 'https://www.tribuneindia.com/archive_search.aspx?txt=Flooding&catid=2&fd=01/01/2010%2012:00:00%20AM&td=11/01/2019%2012:00:00%20AM'
    res = requests.get(url)

    res.status_code
    soup = BeautifulSoup(res.content, 'lxml')
    
    urls = [i('a', href=True)[0]['href'] for i in soup.find_all('div', {'class': 'listing'})]
    headlines = [i('a')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    dates = [i('span')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    captions = [i('p')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    
    d = {'headline': headlines, 
     'date': dates, 
     'caption': captions, 
     'url': urls}
    
    df_mother = pd.DataFrame(d)
    
    time.sleep(2)
    
    for i in range(2,6): 
        
        url1 = url + '&page=' + str(i)
        res = requests.get(url1)
        
        print(res.status_code)
        
        soup = BeautifulSoup(res.content, 'lxml')
        
        urls = [i('a', href=True)[0]['href'] for i in soup.find_all('div', {'class': 'listing'})]
        headlines = [i('a')[0].text for i in soup.find_all('div', {'class': 'listing'})]
        dates = [i('span')[0].text for i in soup.find_all('div', {'class': 'listing'})]
        captions = [i('p')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    
        d2 = {'headline': headlines, 
             'date': dates, 
             'caption': captions, 
             'url': urls}
    
        df = pd.DataFrame(d2)

        print(len(urls))
        print(len(dates))
        
        df_mother = pd.concat([df_mother, df], ignore_index=True)
        
        time.sleep(2)
        
    return(df_mother)
        

url_df = web_df()

print(type(url_df))
print(url_df.shape)

url_df.to_csv('../data/trib_flooding_url.csv', index=False)

df = pd.read_csv("../data/trib_flooding_url.csv")

def content_scrape():
    
    content_dict = {}
    
    for i in range(120):
        url_base = 'https://www.tribuneindia.com/'
        url = url_base + df['url'][i]

        res = requests.get(url)

        print(res.status_code)
        soup = BeautifulSoup(res.content, 'lxml')
        content = [i.text.replace('\n', ' ') for i in soup.find_all('span', {'class': 'storyText'})]
        
        content_dict.update({i: content})
        
        time.sleep(2)
    
    return content_dict
    

content_dict = content_scrape()

len(content_dict)

trib_flooded_text = pd.DataFrame.from_dict(content_dict, orient='index')
trib_flooded_text.rename(columns={0: 'text'}, inplace=True)
trib_flooded_text.to_csv('../data/trib_flooding_text.csv', index=False)

## Heavy Rain

def web_df():

    url = 'https://www.tribuneindia.com/archive_search.aspx?txt=heavy%20rain&catid=2&fd=01/01/2010%2012:00:00%20AM&td=11/01/2019%2012:00:00%20AM'
    res = requests.get(url)

    res.status_code
    soup = BeautifulSoup(res.content, 'lxml')
    
    urls = [i('a', href=True)[0]['href'] for i in soup.find_all('div', {'class': 'listing'})]
    headlines = [i('a')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    dates = [i('span')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    captions = [i('p')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    
    d = {'headline': headlines, 
     'date': dates, 
     'caption': captions, 
     'url': urls}
    
    df_mother = pd.DataFrame(d)
    
    time.sleep(2)
    
    for i in range(2,7): 
        
        url1 = url + '&page=' + str(i)
        res = requests.get(url1)
        
        print(res.status_code)
        
        soup = BeautifulSoup(res.content, 'lxml')
        
        urls = [i('a', href=True)[0]['href'] for i in soup.find_all('div', {'class': 'listing'})]
        headlines = [i('a')[0].text for i in soup.find_all('div', {'class': 'listing'})]
        dates = [i('span')[0].text for i in soup.find_all('div', {'class': 'listing'})]
        captions = [i('p')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    
        d2 = {'headline': headlines, 
             'date': dates, 
             'caption': captions, 
             'url': urls}
    
        df = pd.DataFrame(d2)

        print(len(urls))
        print(len(dates))
        
        df_mother = pd.concat([df_mother, df], ignore_index=True)
        
        time.sleep(2)
        
    return(df_mother)
        

url_df = web_df()

print(type(url_df))
print(url_df.shape)

url_df.to_csv('../data/trib_heavy_rain_url.csv', index=False)

df = pd.read_csv("../data/trib_heavy_rain_url.csv")
print(df.shape)

def content_scrape():
    
    content_dict = {}
    
    for i in range(118):
        url_base = 'https://www.tribuneindia.com/'
        url = url_base + df['url'][i]

        res = requests.get(url)

        print(res.status_code)
        soup = BeautifulSoup(res.content, 'lxml')
        content = [i.text.replace('\n', ' ') for i in soup.find_all('span', {'class': 'storyText'})]
        
        content_dict.update({i: content})
        
        time.sleep(2)
    
    return content_dict

content_dict = content_scrape()

trib_heavy_rain_text = pd.DataFrame.from_dict(content_dict, orient='index')
trib_heavy_rain_text.rename(columns={0: 'text'}, inplace=True)
trib_heavy_rain_text.drop(columns=[1,2,3,4], inplace=True)

trib_heavy_rain_text.to_csv('../data/trib_heavy_rain_text.csv', index=False)

## Monsoon

def web_df():

    url = 'https://www.tribuneindia.com/archive_search.aspx?txt=Monsoon&catid=2&fd=01/01/2010%2012:00:00%20AM&td=11/01/2019%2012:00:00%20AM'
    res = requests.get(url)

    res.status_code
    soup = BeautifulSoup(res.content, 'lxml')
    
    urls = [i('a', href=True)[0]['href'] for i in soup.find_all('div', {'class': 'listing'})]
    headlines = [i('a')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    dates = [i('span')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    captions = [i('p')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    
    d = {'headline': headlines, 
     'date': dates, 
     'caption': captions, 
     'url': urls}
    
    df_mother = pd.DataFrame(d)
    
    time.sleep(2)
    
    for i in range(2,18): 
        
        url1 = url + '&page=' + str(i)
        res = requests.get(url1)
        
        time.sleep(5)
        
        print(res.status_code)
        
        soup = BeautifulSoup(res.content, 'lxml')
        
        urls = [i('a', href=True)[0]['href'] for i in soup.find_all('div', {'class': 'listing'})]
        headlines = [i('a')[0].text for i in soup.find_all('div', {'class': 'listing'})]
        dates = [i('span')[0].text for i in soup.find_all('div', {'class': 'listing'})]
        captions = [i('p')[0].text for i in soup.find_all('div', {'class': 'listing'})]
    
        d2 = {'headline': headlines, 
             'date': dates, 
             'caption': captions, 
             'url': urls}
    
        df = pd.DataFrame(d2)

        print(len(urls))
        print(len(dates))
        
        df_mother = pd.concat([df_mother, df], ignore_index=True)
        
        time.sleep(2)
    
    return(df_mother)
        

url_df = web_df()

print(type(url_df))
print(url_df.shape)

url_df.to_csv('../data/trib_monsoon_url.csv', index=False)

df = pd.read_csv("../data/trib_monsoon_url.csv")

def content_scrape():
    
    content_dict = {}
    
    for i in range(323):
        url_base = 'https://www.tribuneindia.com/'
        url = url_base + df['url'][i]

        res = requests.get(url)

        print(res.status_code)
        soup = BeautifulSoup(res.content, 'lxml')
        content = [i.text.replace('\n', ' ') for i in soup.find_all('span', {'class': 'storyText'})]
        
        content_dict.update({i: content})
        
        time.sleep(3)
    
    return content_dict
    

content_dict = content_scrape()

trib_monsoon_text = pd.DataFrame.from_dict(content_dict, orient='index')
trib_monsoon_text.drop(columns=[1], inplace=True)
trib_monsoon_text.rename(columns={0:'text'}, inplace=True)
trib_monsoon_text.to_csv('../data/trib_monsoon_text.csv', index=False)
