==> News Sources - Gathering, Cleaning, Combining/02 Cleaning and Preparing_Tribune.ipynb <==
import pandas as pd
import numpy as np
import re

import matplotlib.pyplot as plt
%matplotlib inline

## Read csv files

flood_text = pd.read_csv('../data/trib_flood_text.csv')

flooded_text = pd.read_csv('../data/trib_flooded_text.csv')

flooding_text = pd.read_csv('../data/trib_flooding_text.csv')

monsoon_text = pd.read_csv('../data/trib_monsoon_text.csv')

heavy_rain_text = pd.read_csv('../data/trib_heavy_rain_text.csv')

flood_url = pd.read_csv('../data/trib_flood_url.csv')

flooded_url = pd.read_csv('../data/trib_flooded_url.csv')

flooding_url = pd.read_csv('../data/trib_flooding_url.csv')

monsoon_url = pd.read_csv('../data/trib_monsoon_url.csv')

heavy_rain_url = pd.read_csv('../data/trib_heavy_rain_url.csv')

## Formatting

- Make date column datetime object

- Enter full url into url column 

- Format combined csv into 3 columns: date, url and combined text (text)

- Combine headline and text

### *make date a datetime object*

url_list = [flood_url, flooded_url, flooding_url, monsoon_url, heavy_rain_url]

def to_datetime(df_list):
    
    for df in df_list:  
        #removing time stamp
        df['date']= df['date'].map(lambda x: x[:-10])
        
        #converting to datetime object
        df['date'] = pd.to_datetime(df['date'])

to_datetime(url_list)

### *check to see if date changed to datetime*

def datatype(df_list):
    for df in df_list:
        print(df.info())
        print('')
        
datatype(url_list)

### *add full url to url column*

flood_url['url'] = flood_url['url'].map(lambda x: 'https://www.tribuneindia.com' + x)
flooded_url['url'] = flooded_url['url'].map(lambda x: 'https://www.tribuneindia.com' + x)
flooding_url['url'] = flooding_url['url'].map(lambda x: 'https://www.tribuneindia.com' + x)
monsoon_url['url'] = monsoon_url['url'].map(lambda x: 'https://www.tribuneindia.com' + x)
heavy_rain_url['url'] = heavy_rain_url['url'].map(lambda x: 'https://www.tribuneindia.com' + x)

### *scaled df*

flood_url = flood_url[['date', 'url', 'headline']]
flooded_url = flooded_url[['date', 'url', 'headline']]
flooding_url = flooding_url[['date', 'url', 'headline']]
monsoon_url = monsoon_url[['date', 'url', 'headline']]
heavy_rain_url = heavy_rain_url[['date', 'url', 'headline']]

flood_url.head()

### *Combine headline from df_url with text from df_text*

flood = flood_url.join(flood_text)
flooded = flooded_url.join(flooded_text)
flooding = flooding_url.join(flooding_text)
monsoon = monsoon_url.join(monsoon_text)
heavy_rain = heavy_rain_url.join(heavy_rain_text)

df_list = [flood, flooded, flooding, monsoon, heavy_rain]

datatype(df_list)

df_list = [flood, flooded, flooding, monsoon, heavy_rain]

for df in df_list: 
    df['text'] = df['headline'] + '. ' + df['text']
    df.drop(columns=['headline'], inplace=True)

## Check null values

def null_vals(df_list):
    for df in df_list:
        print(df.isnull().sum())
        print('')
        
null_vals(df_list)

### *drop null value in monsoon*

#find index of null value
monsoon[monsoon['text'].isnull()].index.tolist()

monsoon.loc[monsoon[monsoon['text'].isnull()].index]

monsoon.shape

monsoon.drop(index=25, inplace=True)

monsoon.shape

### *check datatypes*

datatype(df_list)

## Clean text 

- lower()

- replace line breaks with spaces

- keep punctation -- not done

#adapted from Dominika Twardowski

def text_processing(df):

    # change line breaks into a space
    df['text'] = df['text'].replace(r"[\r\n]+", " ", regex=True)
    
    # change text to lower case  
    df['text'] = df['text'].map(lambda x: x.lower())
    

for df in df_list: 
    text_processing(df)

heavy_rain.text

## Combine dataframes

tribune_data = pd.concat(df_list, ignore_index=True)

tribune_data.isnull().sum()

tribune_data.rename(columns={'url': 'source'}, inplace=True)

tribune_data.to_csv('../data/tribune_data_clean.csv', index=False)
