==> News Sources - Gathering, Cleaning, Combining/02 Cleaning_ReliefWeb.ipynb <==
## Cleaning the ReliefWeb textual data
---
**Purpose**: This notebook will show the steps which will be followed to clean the collected data from the webscrapping. The data should be in a csv format and the column referring to the date of publication should have been changed to a datetime type OR a datetime type column should have been appended to the csv file (this was accomplished in the previous notebook). The bare minimum of columns required for each csv file should be: date (in datetime type), title (object type), and the text from the article (object type). Later on the source, where the webscrapping was done, will be added to the file.  

This notebook will prepare a csv file for eventual merging with others for eventual analysis and data extraction. 

# Libraries 
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import re

%matplotlib inline

## Read in the csv file
Read in the csv file. Take special care to remember to add the argument `parse_dates` with the column name of the column which has the dates in datetime format. This will allow it to be imported as a datetime type instead of an object. This argument will have to be used whenever a datetime type column is part of the csv file. 

# Read in data from reliefweb AFTER 2006
# Notice: after the csv name there is an additional argument 
    
reliefweb06_df= pd.read_csv("../data/reliefweb_floods_after06.csv",
                            parse_dates = ["datetime"])

### Examine the dataframe
Take a quick look at the imported csv file. Check for null values, number of rows, and data types. 

# check
reliefweb06_df.head()

# check types and non-null value count
reliefweb06_df.info()

# double check for nulls
reliefweb06_df.isnull().sum()

After the initial look at the dataframe, I can see there are no null values for any of the rows. There are 344 rows of data pulled from reliefweb.int which are the articles after 2006. All the columns, except the datetime are object (string) types which makes sense for this being textual data. The datetime is a datetime type which was created from the date (string) in the previous notebook.

# check how the datetime data is formatted
reliefweb06_df["datetime"][1]

### Source column 
A column named "source" will be created. This column will show from where the data comes.   Since this entire csv file is from one source (reliefweb.int) the column will be created and "reliefweb" will be inserted for each observation.

# Create column for the source of this data
reliefweb06_df["source"] = "reliefweb"

# check
reliefweb06_df.head()

## Processing the text data
Convert all text to lowercase and replace line breaks with spaces. This will be done for the title and body of each row. 

def text_processing(df, column): 
    # Change all text to lowercase for movies dataframe body 
    df[column]= df[column].map(lambda x: x.lower())

    # Change line breaks into a space
    df[column] = df[column].replace(r"[\r\n]+", " ", regex=True)

# processing the title and body of each row of the dataframe
text_processing(reliefweb06_df, "body")
text_processing(reliefweb06_df, "title")

# check
reliefweb06_df.head()

### Create a column which combines the title and body 
A new column called "text" will be created which is the text from the "title" and "body" of each observation with a space separating them.

# Function to combine columns of text into one new text column for the dataframe
def combine_text(df, column1, column2):
    df["text"] = df[column1] + " " + df[column2]

# combine the title and body of each observation in the dataframe
combine_text(reliefweb06_df, "title", "body")

# check
reliefweb06_df.head()

### Create "new" dataframe 
In preparation of eventually combining all the collected csv files into one, the format  will be a text (combined all text) column,a date (as a datetime type) column, and a source(naming where the data was pulled from) column. 
As a precaution against running into a copy warning, `.copy()` is added to the end of the operation. The copy function will create a copy of the columns of the dataframe and when either of dataframes is modified it will not affect the other dataframe. 

# making new dataframe from the previous one
# select the rows interested in: datetime (the date as datetime type), text, source
reliefweb_new = reliefweb06_df[['datetime', 'url', 'text']].copy()

# check
reliefweb_new.head()

# check the types 
reliefweb_new.info()

### Rename column names 
To help with eventual combining of all the csv files (containing text data related to our database search), certain columns will be renamed to be uniform across all csv files. The columns will be a "date" column which will be a datetime type column of when the article was published, a "source" column which states the source of the data, and a "text" column which has the text data of the article (the title and article body).

# function to rename columns to "date", "source", "text"
def rename_col(df, column_date, column_source, column_text):
    df.rename(columns={column_date: "date",
                       column_source: "source", 
                       column_text :"text"}, 
              inplace = True)

# apply function to the dataframe
rename_col(reliefweb_new, "datetime", "url", "text")

# check
reliefweb_new.head()

## Save the new dataframe to a csv
This "new" dataframe should only have the three expected columns and has all the cleaning done to the text. 

# create csv file from the new dataframe
reliefweb_new.to_csv("../data/reliefweb_clean.csv", index = False)

**Summary**: In this notebook a csv file was cleaned and processed to get it correctly formatted for text analysis. A new csv file was also created including the columns which will be required for the ultimate combined csv of all the csv files. In the next notebook, some intial text analysis will be done to try to find the most important sentences from the articles collected.
