==> News Sources - Gathering, Cleaning, Combining/02a Translating and Formatting _Dainik.ipynb <==
import pandas as pd
import numpy as np
from bs4 import BeautifulSoup 
import time
from textblob import TextBlob, Word
from googletrans import Translator
import re

pip install googletrans

df = pd.read_csv("../data/flood_payal.csv")

df.head()

df.drop("Unnamed: 0",axis=1,inplace=True)

def headlines_to_words(raw_review):
    # Function to convert a raw review to a string of words
    # The input is a single string (a raw movie review), and 
    # the output is a single string (a preprocessed movie review)
    
    # 1. Remove HTML.
    review_text = BeautifulSoup(raw_review).get_text()
    
# #     2. Remove non-letters.
#     letters_only = re.sub("[^a-zA-Z0-9]", " ", review_text)
    
    # 3. Convert to lower case, split into individual words.
    words = review_text.lower().split()
    
#     # 4. In Python, searching a "set" is much faster than searching
#     # a list, so convert the stop words to a "set".
#     stops = set(stopwords.words('english'))
    
#     # 5. Remove stop words.
#     meaningful_words = [w for w in words if w not in stops]
    #translate into english:
    
    
    # 6. Join the words back into one string separated by space, 
    # and return the result.
    return(" ".join(words))

for i in range(len(df['heading'])):
  
    df['heading'][i]=BeautifulSoup(df['heading'][i]).get_text()

# for i in range(len(df['date'])):
  
#     df['date'][i]=BeautifulSoup(df['date'][i]).get_text()

for i in range(len(df['heading'])):
    df['heading'][i]=df['heading'][i].replace("[,","").replace("[","").replace("।","").replace(",","").replace("(","").replace(":","").replace(")","").replace("\n","").replace("|","").replace(".","")

# for i in range(len(df['date'])):
#     df['date'][i]=df['date'][i].replace("[Updated:","").replace("[","").replace("-","").replace(",","").replace("(","").replace(":","").replace("/n","").replace("IST First Published\r\n\t\t\t\t\t\t\t\t\t","").replace("Posted On ","").replace("]","")

# df['date'][21] ='Dec 31 2015 2311 IST'

# df['date'][21]

# translator = Translator()

# #     df['heading'][i] = translator.translate(df['heading'][i],dest='en',src='hi')
    
# txt = TextBlob(df['heading'][30]).translate(to = 'en')


# txt

# df['heading'][30]="Indo-Pak borders washed away due to floods"

translated =['The scene of the bus stand chowk was inundated after the rain in Mansa on Wednesday  Nis Budhlada 29 August Nis today Indra Devta came to Manasa for the second time and heavy rains submerged the city completely Weather suddenly changed the weather I had a cold but also created problems for the people While people enjoyed the rain the streets of the city were flooded with water and people left their homes It was seen pouring water from it it was raining a bit since morning and after noon the entire city was submerged due to rain the entire city was inundated Gyan Shahar Bus Stand Chowk Rambagh Road Link Road Kachhari Road Cinema Road Veer Nagar Mohalla Gaushala Road Hospital Road Water Curves Road and Jawaharke Road the streets and streets of the city remained waterlogged throughout the day the streets and roads in the lower areas of the city were inundated by rain water Due to the absence of water drainage people faced a lot of difficulty even the vehicles passing through the water stored on the roads were stopped in the middle in many places due to excess water people were also seen attacking the vehicles Due to the huge amount of water accumulated on the road including Cinema Road shopkeepers had to face two problems due to various problems Rain water also entered inside many shops Apart from other areas of Hara people of ward number 1 10 and 8 had a lot of trouble because in the absence of water drainage where there is always water on the roads due to several hours of rain water logging increased further Number nine has also received reports of water logging in many houses; on the other hand three wheelers and rickshaw drivers have made a lot of profit due to the rain while the road drivers have got a lot due to rain Reshan was like this was in Kachari road and bus stand chowk where people faced lot of difficulties in coming due to accumulation of water besides Rambagh Road and underbridge water of the city including water in other low areas Due to the problems faced by the people the streets were flooded till late evening due to which the passers got problems',
            "Hoshiarpur July 31 The roof of the house of a poor family collapsed due to heavy rains in Nisak village Sakrooli last night but the house has been damaged but according to the information the victim Ram Kishore along with his family was found in the house last night When the rain came at a high speed around 2 clock in the night suddenly the roof of the house fell coincidentally the roof of the house fell on the other side all the members slept inside the house Bengal Child escaped but collapsed house had inside the debris of valuables roof",
            "The flood gates of the Bhakra Dam were opened on Saturday after the Tribune News Service flooded the village bastion  Daily Tribune / Manoj Mahajan New Delhi Chandigarh 21 August Heavy rain for the last few hours caused more water to be released from flood dams in the river drains In northern India life has become a mess in places Delhi Punjab Haryana Himachal Chandigarh Uttarakhand Jammu and Kashmir and Rajasthan and North The flooding of rain in many places in the country has increased the hardships of people and alerted in many places due to overflowing of Bhakra Dam increased water flow in Sutlej and Yamuna increased the vigil in the areas along the river Has gone and people have been sent to safe places due to the filling of rain water in many places in the national capital the problems of the people have increased The water level in Yamuna has crossed the danger mark due to rain the government has shifted the people living in the low lying areas along the river to safe places Last evening the Yamuna water level has crossed the danger mark of 20483 meters The cm is flowing up people living in some lowlying areas of the eastern part of the capital have been shifted to safe places where the meteorological department has shifted the northwest Heavy rainfall has been predicted in some places during the next twentyfour hours in the state of Chandigarh Heavy rains occurred in Chandigarh last night causing problems of waterlogging in the lowlying areas causing power failure in many areas and water in many colonies to the squares But on the knees there was water due to which people had to face a total of 869 mm of rain in the city in the coming hours in Haryana Heavy rains are expected in most places in Punjab there is a possibility of heavy rains in many places in Punjab Due to the release of water from barrage in Haryana and rain in Himachal and Shiwaliks hills Yamuna Tangri Markanda Rune and Begna rivers have no water level No improvement has been made in flood situation in Sonipat and Delhi in Yamuna Necessary preparations have been completed People living in and around Yamuna have been evacuated to safer places There has been widespread rainfall in Haryana 65 mm Hisar 35 mm Gurgaon 12 mm Narnaul 14 mm Rohtak 56 mm Bhiwani 14 mm Kalka 20 mm Panchkula in Ambala 15 mm Morni 15 and many other places received rain and dense cloud cover in the state Nangal Dam in Punjab 604 mm Shahpur Kandi 296 mm Roped 43 mm Gurdaspur district received heavy rainfall Ranjit Sagar Dam 584 mm and RL 1700 25 mm rainfall Bhuntar 12 mm Shimla 259 mm Sundernagar 99 mm Kalpa 64 mm Pandoh 40 mm Rampur 132 mm Sunni 305 mm Kasol 71 mm Sujanpur 12 mm Due to heavy rains in Uttarakhand the Tehri dam has started filling up due to increase in the water level of the rivers the dam has also recorded the highest power generation so far According to the top official of the Tehri Hydropower Development Corporation the water level of this largest artificial reservoir in Asia has gone up from 8165 meters and its 820 meter filling limit is coming soon It should release 200 cusecs of water daily It has been the release of so much water in the Ganges that a flood situation can occur in the plains",
            ]

for i in range(0,3):
    df['heading'][i]=translated[i]

df

# df['heading'][24]="'Flood on the border, BSF outpost disconnected from the country pathankot bureau"

df.to_csv("../data/clean_flood1.csv")
