==> News Sources - Gathering, Cleaning, Combining/03a Combining_Data.ipynb <==
# Combining the Datasources
---
**Purpose**: Combining all the important sentences from the various news sources we had into one dataframe of important sentences. Then I will make dataframes related to different important categories (such as: lives lost, humans affected, economic affects).

# Libraries 
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import re

%matplotlib inline

### Read in the csv files
Read in the csv files of the important sentences. Notice the additional argument called `parse_dates` set equal to the column "date" which has the publication date of the article in datetime format. 

# Read in data from reliefweb AFTER 2006
# Notice: after the csv name there is an additional argument 
    
reliefweb_df= pd.read_csv("../data/reliefweb_clean.csv",
                            parse_dates = ["date"])

tribune_df = pd.read_csv("../data/tribune_data_clean.csv",
            parse_dates = ["date"])

times_df= pd.read_csv("../data/payal_clean.csv",
            parse_dates = ["date"])

# list of all the dataframes which will be combined together
dataframes = [tribune_df, times_df, reliefweb_df]

# combined dataframe
combined_df = pd.concat(dataframes, ignore_index=True)

combined_df.head()

combined_df.info()

## Find duplicated articles
Since the same article could be published from a different website or the same article could be published on different days with a different url. I wanted to make sure that we did not have duplicated information going forward. 

# Find duplicated articles based on duplicated text 
dups = combined_df["text"]
combined_df[dups.isin(dups[dups.duplicated()])].sort_values("date")

# Keeping the first appearance of a duplicated text
combined_df = combined_df.drop_duplicates(subset = "text", keep = "first")
combined_df = combined_df.reset_index()

# check 
dups = combined_df["text"]
combined_df[dups.isin(dups[dups.duplicated()])].sort_values("date")

A little over 200 articles were repeated and they were removed from the dataframe. 

# check info
combined_df.info()

# save to csv file 
# create csv file from the new dataframe
combined_df.to_csv("../data/combined.csv", index = False)

**Summary**: In this notebook, the cleaned data from each of the news source were combined together into one dataframe and then saved as a single csv. Next this csv file will be used to extract meaningful information from it to make a database about historic floods in Punjab, India.
