==> News Sources - Gathering, Cleaning, Combining/03b Analysis of Text.ipynb <==
# Analysis of Text 

**Purpose:** In this notebook the individual keyword searches are combined into one dataframe with an additional column which is related to what keyword search the row was created from. 


# Libraries 
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import re

%matplotlib inline

# Read in the new csv file
# Notice: additional  parse_dates argument     
combined_df= pd.read_csv("../data/combined.csv",
                            parse_dates = ["date"])

combined_df.info()

# check if have any duplicated articles 
dups = combined_df["text"]
combined_df[dups.isin(dups[dups.duplicated()])].sort_values("text")

# function to make dataframe from search_keywords
def imp_sentences(dataframe, search_keywords):

    sentences_list = []
    
    numeric_words = [" zero","one ", " two", " three", " four", " five", " six", 
                 "seven", "eight", "nine", " ten", "eleven", "twelve", "thirteen", "fourteen",
                 "fifteen", "sixteen", "seventeen", "eighteen", "nineteen","twenty",
                 "thirty", "forty", "fifty", "sixty", "seventy", "eighty", "ninety",
                 "hundred", "thousand", "million", "billion", "trillion"]


    # function for finding string with a number in it
    def hasnumber(inputsentence):
        return bool(re.search(r'\d', inputsentence))
    
    # function for finding string with numeric word in it
    def hasnumword(inputsentence):
        has_numericword = False
        for keyword in numeric_words:
            if len(re.findall(keyword,inputsentence))>0:
                has_numericword = True
        return has_numericword

    # loop through the entire dataframe
    for idx in range(0,dataframe.shape[0]):
        row = dataframe.iloc[idx,:]

        # split the row on ". " 
        # sentences = row["text"].split(". " )
        sentences = re.split(r'\.\D', row["text"])
        date = row["date"]
        source = row["source"]

        # for each sentence of the row
        for sentence in sentences:
            has_keyword = False
            for keyword in search_keywords:
                #if sentence has keyword:
                # length will be greater than zero
                if len(re.findall(keyword, sentence))>0:
                     #set has_keyword to True
                    has_keyword=True
                     #break out of the loop
                    break

            number_bool = hasnumber(sentence) 
            has_numericword = hasnumword(sentence)

            # if the sentence has at least one of the keywords
            if (has_keyword == True and number_bool == True) or (has_keyword == True and has_numericword == True) :
                sentence_info = {}
                sentence_info["date"] = date
                sentence_info["source"] = source
                sentence_info['sentence'] = sentence
                # add the sentence and the date to the sentence list
                sentences_list.append(sentence_info)  
                
    search_keywords_df =  pd.DataFrame(sentences_list)
    return search_keywords_df

# keywords

total_keywords= ['crop', 'livestock', 'cattle', "farm", 'wheat', 'rice', 'barley',
                 'agricult', 'acre',
                'death', ' die', 'dead','kill', 'drown', "mortality",
                'victim', 'suffer', 'rescue', 'missing', 'evacuat', 'help', 'affect',
                'relief', 'wound', 'survive', 'displace', 'strand','lost', 'damag',
                'destroy', 'econom', 'cost', 'rehabilit'
                'house', 'home', 'roof', 'street', 'road', 'bridge']
# find total search_keywords
imp_sentences(combined_df, total_keywords)

total_keyword_one_search_df= imp_sentences(combined_df, total_keywords)

total_keyword_one_search_df.info()

# Find duplicated sentences based on duplicated text  in the total_keyword search
dups = total_keyword_one_search_df["sentence"]
total_keyword_one_search_df[dups.isin(dups[dups.duplicated()])].sort_values("sentence")

#keywords
agri_keywords=['crop', 'livestock', 'cattle', "farm", 'wheat', 'rice', 'barley',
              'agricult', 'acre'
              ]

# find agriculture data 
imp_sentences(combined_df, agri_keywords)

agri_df = imp_sentences(combined_df, agri_keywords)

# add column named to specify what type it is 
agri_df["topic"] = "agri"

# check
agri_df.head()

# keywords 
death_keywords=['death', ' die', 'dead','kill', 'drown', "mortality"]

# find sentences related to death
imp_sentences(combined_df, death_keywords)

# create dataframe 
death_df = imp_sentences(combined_df, death_keywords)

# add column named to specify what type it is 
death_df["topic"] = "death"

# check
death_df.head()

death_df.loc[100].sentence

# keywords
lives_keywords = ['victim', 'suffer', 'rescue', 'missing', 'evacuat', 'help', 'affect',
                  'relief', 'wound', 'survive', 'displace', 'strand','lost', 'damag',
                  'destroy', 'econom', 'cost', 'rehabilit',
                 ]

# find the lives data 
imp_sentences(combined_df, lives_keywords)

# create dataframe 
lives_df = imp_sentences(combined_df, lives_keywords)

# add column named to specify what type it is 
lives_df["topic"] = "lives"

# check
lives_df.head()

lives_df.loc[30].sentence

# keywords house/infastructure
house_keywords = [ 'house', 'home', 'roof', 'street', 'road', 'bridge', ]

imp_sentences(combined_df, house_keywords)

# create dataframe 
houses_df = imp_sentences(combined_df, house_keywords)

# add column named to specify what type it is 
houses_df["topic"] = "house"

# check
houses_df.head()

houses_df.loc[5].sentence

### combine the dataframe topics together into one dataframe 

# list of all the dataframes which will be combined together
dataframes = [death_df, lives_df, agri_df, houses_df]

# combined dataframe
keyword_concat_df = pd.concat(dataframes, ignore_index=True)

keyword_concat_df.info()

total_keyword_one_search_df.info()

## combining all the keyword lists into separate areas adds on a little over 1000 rows

total_keyword_one_search_df.loc[20].sentence

### find duplicated rows in the combined dataframe 

keyword_concat_df[keyword_concat_df.duplicated(keep = False)].sort_values("source")

There are 67 duplicate rows. The duplicates make sense for the reliefweb articles since these articles report on multiple states and locations in one article and they repeat the sentence "no human death has been reported in ..." if there was no death in a location. 

# Keeping the first appearance of a duplicated text
keyword_concat_df = keyword_concat_df.drop_duplicates( keep = "first")
keyword_concat_df = keyword_concat_df.reset_index()

# Reset the index
keyword_concat_df = keyword_concat_df.reset_index()

# check that dropped all duplicated rows
keyword_concat_df[keyword_concat_df.duplicated()]

keyword_concat_df.info()

Now we just keep the first appearance of a duplicate row. Now the dataframe is 4213 entries.

## save to csv 

# save to csv file 
# create csv file from the new dataframe

# dataframe of the individual searches combined together with a topic column!
keyword_concat_df.to_csv("../data/keyword_topic_sentences.csv", index = False)

keyword_concat_df.head()

**Summary**: One large topic concatenated dataframe was created and it has approximately 1000 more rows than a dataframe just passing through the combined keyword search. This means that that some of the sentences are being selected in multiple keyword searches. Next, this csv file will be used to create a SQL database and visuals to represent the findings.
