==> News Sources - Gathering, Cleaning, Combining/05 Finding Location.ipynb <==
## Finding Location from textual data

# Libraries 
import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import re

%matplotlib inline

## Read in the csv files
Read in the combined csv file.

# Read in data from reliefweb AFTER 2006
# Notice: after the csv name there is an additional argument 
    
df= pd.read_csv("./combined.csv")

# check
df.head()

# Import Districts
location= pd.read_csv("./latlongpunjab.csv")
location = location['Location']

location = location.str.lower()

#adding some replacements based on more commonly used place names
# fatehgarh sahib = fatehgarh
# sbs nagar = nawanshahr, sbs
# sas nagar = mohali, nagar
# sas nagar = ajitgarh, nagar 
# sri muktsar sahib = muktsar
added_names = pd.Series(['fatehgarh','nawanshahr','mohali','ajitgarh','muktsar','nagar','sbs'])

location = location.append(added_names)

# Finding Location in the Text & URLs

# Write a function to pass in the list
def location_fcn(lst, df):
    for i in lst:
        df[i] = df['source'].str.count(i) + df['text'].str.count(i)

location_fcn(location, df)

df.sum()

#combining columns for the extra names back to the original name
# fatehgarh sahib = fatehgarh
# sbs nagar = nawanshahr, sbs
# sas nagar = mohali, nagar
# sas nagar = ajitgarh, nagar 
# sri muktsar sahib = muktsar
df['fatehgarh sahib'] = df['fatehgarh sahib'] + df['fatehgarh']
df['sbs nagar'] = df['sbs nagar'] + df['nawanshahr']+ df['sbs']
df['sas nagar'] = df['sas nagar'] + df['mohali']+ df['nagar']+ df['ajitgarh']
df['sri muktsar sahib'] = df['sri muktsar sahib'] + df['muktsar']

df = df.drop(['muktsar','mohali','nagar','ajitgarh','nawanshahr','sbs','fatehgarh'], axis = 1)

df.sum()

# Finding People, Crops, Homes, Money in the Text & URLs

# Finding People Affected
def people_fcn(lst, df):
    sum_people = 0
    for i in lst:
        sum_people += df['source'].str.count(i) + df['text'].str.count(i)
    df['people'] = sum_people

peeps = ['death', 'die', 'dead', 'people','resident','residents',
                 'kill', 'evacuat', 'mortality','villager','villagers',
                 'victim', 'suffer', 'rescue', 'toll', 'displace','displaced',
                 'missing', 'drown', 'wound', 'survive','persons','family','families',
                 'suffer', 'displace', 'strand', 'displace','person']

people_fcn(peeps, df)

df['people'].value_counts()

#Finding Crops / Livestock Affected
crop_livestock = ['damag', 'loss', 'lost','kill','farm','field','agriculture','stable','horse','horses',
                 'destroy','crop', 'livestock', 'cattle','farmer','agricultural','land','lands','fields',
                  'farmers','acre','acres','grain','basmati','paddy',
                 'cow','sheep','pig','cows','pigs','corn','wheat','seed','soy','rice', 'crops']

# Finding Crops / Livestock Affected
def crops_fcn(lst, df):
    sum_crops = 0
    for i in lst:
        sum_crops += df['source'].str.count(i) + df['text'].str.count(i)
    df['crops'] = sum_crops


crops_fcn(crop_livestock, df)

df['crops'].value_counts()

#Finding Houses Affected
house = ['damag', 'loss', 'lost', 'evacuat','houses','submerged','rooftop','rooftops','submerge',
                 'affect', 'relief', 'relieve', 'destroy', 'dwelling','structural','village','villages',
                 'house', 'roof', 'home', 'window','damage', 'structure', 'barn','hut']

# Finding Homes Affected
def home_fcn(lst, df):
    sum_homes = 0
    for i in lst:
        sum_homes += df['source'].str.count(i) + df['text'].str.count(i)
    df['homes'] = sum_homes

home_fcn(house, df)

df['homes'].value_counts()

#Finding the Money
money = ['cost','money','dollar','rupee','economic','economy','value','pounds','euro','compensation', 'rupees','dollars','euros']

# Finding the Money
def money_fcn(lst, df):
    sum_money = 0
    for i in lst:
        sum_money += df['source'].str.count(i) + df['text'].str.count(i)
    df['money'] = sum_money

money_fcn(money, df)

df['money'].value_counts()

df.sum()

import pyparsing as pp
import calendar

months_list= []
for month_idx in range(1, 13):
    months_list.append(calendar.month_name[month_idx])
    months_list.append(calendar.month_abbr[month_idx])

months_list = [x.lower() for x in months_list]
# join the list to use it as pyparsing keyword
month_keywords = " ".join(months_list)

# date separator - can be one of '/', '.','-', or ' '
separator = pp.Word("/.,- ")

# Dictionary for numeric date e.g. 12/12/2018
numeric_date = pp.Combine(pp.Word(pp.nums, max=2) + separator + pp.Word(pp.nums, max=2) + separator + pp.Word(pp.nums, max=4))

# Dictionary for text date e.g. 12/Jan/2018
text_date = pp.Combine(pp.Word(pp.nums, max=2) + separator + pp.oneOf(month_keywords) + separator + pp.Word(pp.nums, max=4))
#jan 12, 2018
text_date2 = pp.Combine(pp.oneOf(month_keywords) + separator + pp.Word(pp.nums, max=2) + separator + pp.Word(pp.nums, max=4))

#jan 12
text_date3 = pp.Combine(pp.oneOf(month_keywords) + separator + pp.Word(pp.nums, max=2))

#12 jan
text_date4 = pp.Combine(pp.Word(pp.nums, max=2) + separator + pp.oneOf(month_keywords))

# Finding any type of date - except numeric
date_patterns = [text_date, text_date2, text_date3, text_date4]


#Finding the first date in the text - START HERE CAPTURING DATE
df_dates = {'index_key': [],
            'event_date': []}
date_list = []
# for j in range(0,len(df)):  
for j in range(0,len(df)): 
    for i, date_pattern in enumerate(date_patterns):
        pattern = pp.Suppress(pp.SkipTo(date_pattern)) + date_pattern
        try:
            result = pattern.parseString(df['text'][j])
            df_dates['index_key'].append(j)
            df_dates['event_date'].append(result[0])
#             print(j, result)
        except:
            pass
            

event_dates = pd.DataFrame(df_dates)

# Merging event dates with the dataframe
df3 = pd.merge(df, event_dates, right_on =  "index_key", left_index = True, how="left")

df3['event_date'].fillna("None Found", inplace=True)

#Dropping index key row - not needed now that we've merged
df3.drop("index_key", axis=1, inplace=True)

#Exporting to CSV
df3.to_csv("./location_topic_counts_events.csv", index = False)
